+++ backend/package.json
@@ -12,10 +12,11 @@
   "description": "",
   "dependencies": {
     "@prisma/client": "^6.14.0",
+    "@react-oauth/google": "^0.12.2",
     "bcrypt": "^6.0.0",
     "bcryptjs": "^3.0.2",
     "cors": "^2.8.5",
-    "dotenv": "^17.2.1",
+    "dotenv": "^17.2.2",
     "express": "^5.1.0",
     "jsonwebtoken": "^9.0.2",
     "passport": "^0.7.0",
