+++ backend/src/controllers/authController.js
@@ -1,3 +1,4 @@
+require('dotenv').config();
 const prisma = require('../generated/prisma');
 const jwt = require('jsonwebtoken');
 const bcrypt = require('bcryptjs');
