+++ backend/src/controllers/mensagemController.js
@@ -1,57 +1,63 @@
-const mensagensModel = require('../models/mensagemModel');
+const mensagemModel = require('../models/mensagemModel');
 
-const listarMensagens = async (req, res) => {
+const listarMensagensHandler = async (req, res) => {
   try {
     const mensagens = await mensagemModel.listarMensagens();
     res.status(200).json(mensagens);
   } catch (error) {
-    res.status(500).json({ erro: error.message });
+    res.status(500).json({ error: error.message });
   }
 };
 
-const buscarMensagemPorId = async (req, res) => {
+const buscarMensagemPorIdHandler = async (req, res) => {
   try {
     const { id } = req.params;
-    const mensagem = await mensagemModel.buscarMensagemPorId(Number(id));
+    const mensagem = await mensagemModel.buscarMensagemPorId(parseInt(id));
+
+    if (!mensagem) {
+      return res.status(404).json({ error: 'Mensagem não encontrada' });
+    }
     res.status(200).json(mensagem);
   } catch (error) {
-    res.status(404).json({ erro: error.message });
+    res.status(500).json({ error: error.message });
   }
 };
 
-const criarMensagem = async (req, res) => {
+const criarMensagemHandler = async (req, res) => {
   try {
-    const novaMensagem = await mensagemModel.criarMensagem(req.body);
+    const dadosMensagem = req.body;
+    const novaMensagem = await mensagemModel.criarMensagem(dadosMensagem);
     res.status(201).json(novaMensagem);
   } catch (error) {
-    res.status(400).json({ erro: error.message });
+    res.status(400).json({ error: error.message });
   }
 };
 
-const atualizarMensagem = async (req, res) => {
+const atualizarMensagemHandler = async (req, res) => {
   try {
     const { id } = req.params;
-    const mensagemAtualizada = await mensagemModel.atualizarMensagem(Number(id), req.body);
+    const dadosParaAtualizar = req.body;
+    const mensagemAtualizada = await mensagemModel.atualizarMensagem(parseInt(id), dadosParaAtualizar);
     res.status(200).json(mensagemAtualizada);
   } catch (error) {
-    res.status(400).json({ erro: error.message });
+    res.status(404).json({ error: error.message });
   }
 };
 
-const deletarMensagem = async (req, res) => {
+const deletarMensagemHandler = async (req, res) => {
   try {
     const { id } = req.params;
-    await mensagemModel.deletarMensagem(Number(id));
+    await mensagemModel.deletarMensagem(parseInt(id));
     res.status(204).send();
   } catch (error) {
-    res.status(404).json({ erro: error.message });
+    res.status(404).json({ error: error.message });
   }
 };
 
 module.exports = {
-  listarMensagens,
-  buscarMensagemPorId,
-  criarMensagem,
-  atualizarMensagem,
-  deletarMensagem
+  listarMensagensHandler,
+  buscarMensagemPorIdHandler,
+  criarMensagemHandler,
+  atualizarMensagemHandler,
+  deletarMensagemHandler
 };
