+++ backend/src/controllers/telefoneController.js
@@ -1,41 +1,48 @@
-const prisma = require('../generated/prisma');
+const telefoneModel = require('../models/telefoneModel');
 
-const adicionarTelefone = async (pessoaId, dadosTelefone) => {
-  const { numero, tipo } = dadosTelefone;
-  if (!numero || !tipo) throw new Error('Número e tipo são obrigatórios.');
-
-  const pessoaExistente = await prisma.pessoa.findUnique({ where: { id: pessoaId } });
-  if (!pessoaExistente) throw new Error('Pessoa não encontrada.');
-
-  return prisma.telefone.create({
-    data: { numero, tipo, pessoaId },
-  });
+const adicionarTelefoneHandler = async (req, res) => {
+  try {
+    const { pessoaId } = req.params;
+    const novoTelefone = await telefoneModel.adicionarTelefone(parseInt(pessoaId), req.body);
+    res.status(201).json(novoTelefone);
+  } catch (error) {
+    res.status(400).json({ error: error.message });
+  }
 };
 
-const listarTelefonesPorPessoa = async (pessoaId) => {
-  return prisma.telefone.findMany({ where: { pessoaId: pessoaId } });
+const listarTelefonesHandler = async (req, res) => {
+  try {
+    const { pessoaId } = req.params;
+    const telefones = await telefoneModel.listarTelefonesPorPessoa(parseInt(pessoaId));
+    res.status(200).json(telefones);
+  } catch (error) {
+    res.status(500).json({ error: error.message });
+  }
 };
 
-const atualizarTelefone = async (telefoneId, dadosParaAtualizar) => {
-  const telefoneExistente = await prisma.telefone.findUnique({ where: { id: telefoneId } });
-  if (!telefoneExistente) throw new Error('Telefone não encontrado!');
-
-  return prisma.telefone.update({
-    where: { id: telefoneId },
-    data: dadosParaAtualizar,
-  });
+const atualizarTelefoneHandler = async (req, res) => {
+  try {
+    const { telefoneId } = req.params;
+    const telefoneAtualizado = await telefoneModel.atualizarTelefone(parseInt(telefoneId), req.body);
+    res.status(200).json(telefoneAtualizado);
+  } catch (error) {
+    res.status(404).json({ error: error.message });
+  }
 };
 
-const deletarTelefone = async (telefoneId) => {
-  const telefoneExistente = await prisma.telefone.findUnique({ where: { id: telefoneId } });
-  if (!telefoneExistente) throw new Error('Telefone não encontrado!');
-
-  return prisma.telefone.delete({ where: { id: telefoneId } });
+const deletarTelefoneHandler = async (req, res) => {
+  try {
+    const { telefoneId } = req.params;
+    await telefoneModel.deletarTelefone(parseInt(telefoneId));
+    res.status(204).send();
+  } catch (error) {
+    res.status(404).json({ error: error.message });
+  }
 };
 
 module.exports = {
-  adicionarTelefone,
-  listarTelefonesPorPessoa,
-  atualizarTelefone,
-  deletarTelefone,
+  adicionarTelefoneHandler,
+  listarTelefonesHandler,
+  atualizarTelefoneHandler,
+  deletarTelefoneHandler,
 };
