+++ backend/src/routes/mensagemRoutes.js
@@ -2,14 +2,14 @@ const express = require('express');
 const router = express.Router();
 const mensagemController = require('../controllers/mensagemController');
 
-router.get('/', mensagemController.listarMensagens);
+router.get('/', mensagemController.listarMensagensHandler);
 
-router.get('/:id', mensagemController.buscarMensagemPorId);
+router.get('/:id', mensagemController.buscarMensagemPorIdHandler);
 
-router.post('/', mensagemController.criarMensagem);
+router.post('/', mensagemController.criarMensagemHandler);
 
-router.put('/:id', mensagemController.atualizarMensagem);
+router.put('/:id', mensagemController.atualizarMensagemHandler);
 
-router.delete('/:id', mensagemController.deletarMensagem);
+router.delete('/:id', mensagemController.deletarMensagemHandler);
 
 module.exports = router;
