+++ backend/src/services/loginGoogleServices.js
@@ -1,3 +1,4 @@
+require('dotenv').config();
 const prisma = require('../generated/prisma');
 const passport = require('passport');
 const GoogleStrategy = require('passport-google-oauth20').Strategy;
@@ -6,7 +7,7 @@ passport.use(
   new GoogleStrategy(
     {
       clientID: process.env.GOOGLE_CLIENT_ID,
-      clientSecret: process.env.GOOGLE_SECRET_ID,
+      clientSecret: process.env.GOOGLE_CLIENT_SECRET,
       callbackURL: process.env.CALLBACK_URL
     },
     async (acessToken, refreshToken, profile, done) => {
